Convert Amazonian Mystery pages to tsx
