+++ src/pages/amazonian mystery/SolvedMystery.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 export const SolvedMystery = () => {
   
     return (
